+++ js/src/wp-seo-admin.js
@@ -1,4 +1,4 @@
-/* global wpseoAdminL10n, ajaxurl, tb_remove, wpseoSelect2Locale */
+/* global wpseoAdminL10n, ajaxurl, wpseoSelect2Locale */
 
 import a11ySpeak from "a11y-speak";
 
@@ -96,12 +96,14 @@ import a11ySpeak from "a11y-speak";
 	}
 
 	/**
-	 * Copies the meta description for the homepage
+	 * Copies the meta description for the homepage.
 	 *
 	 * @returns {void}
 	 */
 	function wpseoCopyHomeMeta() {
-		jQuery( "#og_frontpage_desc" ).val( jQuery( "#meta_description" ).val() );
+		jQuery( "#copy-home-meta-description" ).on( "click", function() {
+			jQuery( "#og_frontpage_desc" ).val( jQuery( "#meta_description" ).val() );
+		} );
 	}
 
 	/**
@@ -121,44 +123,6 @@ import a11ySpeak from "a11y-speak";
 	 * When the hash changes, get the base url from the action and then add the current hash
 	 */
 	jQuery( window ).on( "hashchange", wpseoSetTabHash );
-
-	/**
-	 * Add a Facebook admin for via AJAX.
-	 *
-	 * @returns {void}
-	 */
-	function wpseoAddFbAdmin() {
-		var targetForm = jQuery( "#TB_ajaxContent" );
-
-		jQuery.post(
-			ajaxurl,
-			{
-				_wpnonce: targetForm.find( "input[name=fb_admin_nonce]" ).val(),
-				admin_name: targetForm.find( "input[name=fb_admin_name]" ).val(),
-				admin_id: targetForm.find( "input[name=fb_admin_id]" ).val(),
-				action: "wpseo_add_fb_admin",
-			},
-			function( response ) {
-				var resp = jQuery.parseJSON( response );
-
-				targetForm.find( "p.notice" ).remove();
-
-				switch ( resp.success ) {
-					case 1:
-
-						targetForm.find( "input[type=text]" ).val( "" );
-
-						jQuery( "#user_admin" ).append( resp.html );
-						jQuery( "#connected_fb_admins" ).show();
-						tb_remove();
-						break;
-					case 0 :
-						targetForm.find( ".form-wrap" ).prepend( resp.html );
-						break;
-				}
-			}
-		);
-	}
 
 	/**
 	 * Adds select2 for selected fields.
@@ -227,8 +191,6 @@ import a11ySpeak from "a11y-speak";
 	window.setWPOption = setWPOption;
 	window.wpseoCopyHomeMeta = wpseoCopyHomeMeta;
 	// eslint-disable-next-line
-	window.wpseoAddFbAdmin = wpseoAddFbAdmin;
-	window.wpseo_add_fb_admin = wpseoAddFbAdmin;
 	window.wpseoSetTabHash = wpseoSetTabHash;
 
 	jQuery( document ).ready( function() {
@@ -316,6 +278,7 @@ import a11ySpeak from "a11y-speak";
 			}
 		} );
 
+		wpseoCopyHomeMeta();
 		setInitialActiveTab();
 		initSelect2();
 	} );