+++ tests/inc/options/test-class-wpseo-options.php
@@ -163,6 +163,10 @@ class WPSEO_Options_Test extends WPSEO_UnitTestCase {
 	public function test_make_sure_keys_are_unique_over_options() {
 		$keys = array();
 
+		// Make sure the backfilling is not being done when determining "real" unique option names.
+		remove_all_actions( 'option_wpseo' );
+		remove_all_actions( 'option_wpseo_titles' );
+
 		foreach ( array_keys( WPSEO_Options::$options ) as $option_name ) {
 			$option_keys = array_keys( WPSEO_Options::get_option( $option_name ) );
 			$intersected = array_intersect( $option_keys, $keys );