+++ tests/sitemaps/test-class-wpseo-post-type-sitemap-provider.php
@@ -66,7 +66,7 @@ class WPSEO_Post_Type_Sitemap_Provider_Test extends WPSEO_UnitTestCase {
 
 		add_filter( 'wpseo_exclude_from_sitemap_by_post_ids', array( $this, 'filter_with_output' ) );
 
-		$this->assertEquals( array( 5, 600, 23, 0, 0, 3 ) , $sitemap_provider->get_excluded_posts() );
+		$this->assertEquals( array( 5, 600, 23, 0, 0, 3 ), $sitemap_provider->get_excluded_posts() );
 
 		remove_filter( 'wpseo_exclude_from_sitemap_by_post_ids', array( $this, 'filter_with_output' ) );
 	}
@@ -81,7 +81,7 @@ class WPSEO_Post_Type_Sitemap_Provider_Test extends WPSEO_UnitTestCase {
 
 		add_filter( 'wpseo_exclude_from_sitemap_by_post_ids', array( $this, 'filter_with_invalid_output' ) );
 
-		$this->assertEquals( array() , $sitemap_provider->get_excluded_posts( '1,2,3,4' ) );
+		$this->assertEquals( array(), $sitemap_provider->get_excluded_posts( '1,2,3,4' ) );
 
 		remove_filter( 'wpseo_exclude_from_sitemap_by_post_ids', array( $this, 'filter_with_invalid_output' ) );
 	}
@@ -115,7 +115,7 @@ class WPSEO_Post_Type_Sitemap_Provider_Test extends WPSEO_UnitTestCase {
 		return '';
 	}
 
-	/** 
+	/**
 	 * Tests if external URLs are not being included in the sitemap
 	 *
 	 * @covers WPSEO_Post_Type_Sitemap_Provider::get_url