+++ tests/test-class-wpseo-json-ld.php
@@ -39,7 +39,7 @@ class WPSEO_JSON_LD_Test extends WPSEO_UnitTestCase {
 		$home_url   = WPSEO_Utils::home_url();
 		$search_url = $home_url . '?s={search_term_string}';
 		$json       = wp_json_encode( array(
-			'@context'        => 'http://schema.org',
+			'@context'        => 'https://schema.org',
 			'@type'           => 'WebSite',
 			'@id'             => '#website',
 			'url'             => $home_url,
@@ -70,7 +70,7 @@ class WPSEO_JSON_LD_Test extends WPSEO_UnitTestCase {
 
 		$home_url = WPSEO_Utils::home_url();
 		$json     = wp_json_encode( array(
-			'@context' => 'http://schema.org',
+			'@context' => 'https://schema.org',
 			'@type'    => 'Person',
 			'url'      => $home_url,
 			'sameAs'   => array( $instagram ),
@@ -88,15 +88,15 @@ class WPSEO_JSON_LD_Test extends WPSEO_UnitTestCase {
 	 * @covers WPSEO_JSON_LD::organization_or_person()
 	 */
 	public function test_bad_input() {
-		$name      = 'Joost "Yoast":"de Valk"';
-		$home_url  = WPSEO_Utils::home_url();
+		$name     = 'Joost "Yoast":"de Valk"';
+		$home_url = WPSEO_Utils::home_url();
 
 		WPSEO_Options::set( 'company_or_person', 'person' );
 		WPSEO_Options::set( 'person_name', $name );
 		WPSEO_Options::set( 'instagram_url', 'http://instagram.com:8080/{}yoast' );
 
-		$json = wp_json_encode( array(
-			'@context' => 'http://schema.org',
+		$json     = wp_json_encode( array(
+			'@context' => 'https://schema.org',
 			'@type'    => 'Person',
 			'url'      => $home_url,
 			'sameAs'   => array( 'http://instagram.com:8080/yoast' ), // The {} will be stripped out by saving the option.
@@ -126,7 +126,7 @@ class WPSEO_JSON_LD_Test extends WPSEO_UnitTestCase {
 
 		$home_url = WPSEO_Utils::home_url();
 		$json     = wp_json_encode( array(
-			'@context' => 'http://schema.org',
+			'@context' => 'https://schema.org',
 			'@type'    => 'Organization',
 			'url'      => $home_url,
 			'sameAs'   => array( $facebook, $instagram ),